+++ algorithms/NearestNeighbour/NearestNeighbour.h
@@ -15,7 +15,7 @@ using namespace std;
 
 class NearestNeighbour {
 public:
-    int result = INT_MAX;
+    int result = INT16_MAX;
 
     vector<int> best_way = vector<int>();
     void nearestNeighbour(Node *start, int size, vector<Node*> visited, int actualCost,Node *con,int c);
